Moved togglePlayPauseIcon to the js file
