+++ src/togglePlayPauseIcon.js
@@ -1,0 +1,16 @@
+export const togglePlayPauseIcon = (reset) => {
+  const playIcon = document.querySelector('#play-icon');
+  const pauseIcon = document.querySelector('#pause-icon');
+  if (reset) {
+    // when resetting -> always revert to play icon
+    if (playIcon.classList.contains('hidden')) {
+      playIcon.classList.remove('hidden');
+    }
+    if (!pauseIcon.classList.contains('hidden')) {
+      pauseIcon.classList.add('hidden');
+    }
+  } else {
+    playIcon.classList.toggle('hidden');
+    pauseIcon.classList.toggle('hidden');
+  }
+};
